Restructure email package
