+++ .changeset/2791c65a/changes.json
@@ -1,0 +1,1 @@
+{ "releases": [{ "name": "@keystone-alpha/email", "type": "major" }], "dependents": [] }
