+++ .changeset/2791c65a/changes.md
@@ -1,0 +1,1 @@
+- Export { emailSender } as the API, rather than a default export
